fix: loading config file was flakey
The way we were loading the config file caused the integration tests
to be flakey. This changes to process to check if the config path is an
absolute path, and if is not joins it to the exec's working directory.

To make the golden files work correctly, they are now parsed as a regex
and then matched against the captured stderr. This means we can't update
the stderr files, so that method is removed from the test harness.