fix(newsfile): scan for start comments
The original way `stentor` scanned for the start comment worked so long
as the newline after the comment didn't align just after the end of the
internal read buffer. To fix that, stentor now scans the end of the read
buffer looking for partial matches with the start comment.

Fixes #28